🎨 FontDisplayBox 분리
